requirements.txt
asgiref==3.2.10
Django==3.1
django-jsonfield==1.4.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
